Feature/spot#2

+++ src/main/java/com/ssafy/trip/spot/dto/request/NearbySearchRequest.java
@@ -8,6 +8,6 @@ public class NearbySearchRequest {
     private double maxLatitude;
     private double minLongitude;
     private double maxLongitude;
-    private Long cursor; // 마지막 ID
+    private Integer cursor; // 마지막 ID
     private int limit;
 }
