+++ src/main/java/com/ssafy/trip/spot/mapper/SpotMapper.java
@@ -24,7 +24,7 @@ public interface SpotMapper {
 			@Param("maxLat") double maxLat,
 			@Param("minLng") double minLng,
 			@Param("maxLng") double maxLng,
-			@Param("cursor") Long cursor,
+			@Param("cursor") Integer cursor,
 			@Param("limit") int limit
 	);
 }
